Integer to Double + Constructors

+++ converter/SquareConverter.java
@@ -38,8 +38,8 @@ public class SquareConverter {
     // take special note of converting ObjectId to String
     public static Square toSquare(DBObject doc) {
         Square s = new Square();
-        s.setLatitude((Integer) doc.get("latitude"));
-        s.setLongitude((Integer) doc.get("longitude"));
+        s.setLatitude((Double) doc.get("latitude"));
+        s.setLongitude((Double) doc.get("longitude"));
         s.setNearestAtm((Atm) doc.get("nearestAtm"));
         s.setNearestTransport((Transport) doc.get("nearestAtm"));
         s.setNearestSupermarket((Supermarket) doc.get("nearestAtm"));
