+++ model/Square.java
@@ -13,9 +13,9 @@ public class Square {
     
     private String id;
  
-    private Integer latitude;
+    private Double latitude;
  
-    private Integer longitude;
+    private Double longitude;
     
     private Atm nearestAtm;
     
@@ -32,19 +32,19 @@ public class Square {
         this.id = id;
     }
 
-    public Integer getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Integer latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Integer getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Integer longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
